+++ packages/coding-standards/tests/Unit/CsFixer/OrmJoinColumnRequireNullableFixer/correct/one_to_many.php
@@ -1,0 +1,10 @@
+<?php
+
+class Bar
+{
+    /**
+     * @var \StdObject
+     * @ORM\OneToMany(targetEntity="StdObject")
+     */
+    private $foo1;
+}