+++ packages/coding-standards/tests/Unit/Sniffs/ObjectIsCreatedByFactorySniff/Correct/PostFactory.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\CodingStandards\Sniffs\ObjectIsCreatedByFactorySniff\Correct;
+
+final class PostFactory
+{
+    public function create()
+    {
+        $post = new Post();
+    }
+}