+++ packages/coding-standards/tests/autoload.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+$vendorDir = file_exists(__DIR__ . '/../vendor/autoload.php') ? __DIR__ . '/../vendor/' : __DIR__ . '/../../../vendor/';
+
+require_once $vendorDir . 'autoload.php';
+require_once $vendorDir . 'squizlabs/php_codesniffer/autoload.php';