+++ CHANGELOG.md
@@ -63,6 +63,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 #### Fixed
 - [#281 - monorepo-tools: Fix scripts to work on OS X](https://github.com/shopsys/shopsys/pull/282) [@lukaso]
 
+### [shopsys/coding-standards]
+#### Added
+- [#308 - Sniff for forgotten dumps](https://github.com/shopsys/shopsys/pull/308)
+    - ecs tester for coding standards was added with tests for sniffs and fixers [@TomasVotruba]
+
 ### [shopsys/product-feed-google]
 #### Changed
 - [#286 - Instantiate entity data objects by factories](https://github.com/shopsys/shopsys/pull/286)
@@ -1253,3 +1258,4 @@ That's why is this section formatted differently.
 [@EdoBarnas]: https://github.com/EdoBarnas
 [@DavidKuna]: https://github.com/DavidKuna
 [@lukaso]: https://github.com/lukaso
+[@TomasVotruba]: https://github.com/TomasVotruba